Add some default dashboards
Kibana allows you to do cross-index visualization. This is often nice,
but other times it just gets in your way. So we have dashboards for each
index by itself and a dashboard that queries all indices.

"ER" is for ER data. "OTC" is for OTC data. "All" is for both ER and OTC
data.