Primitive elasticsearch proxying for Kibana
This whitelists the elasticsearch URLs + methods that Kibana can hit.
For example, it only allows GET and POST requests. However, it does
not filter the request body (or request parameters). This means that
users can, for example, POST anything to /:index/_search. It also does
not filter responses from elasticsearch at all.