i18n with angular-gettext
The locale is specified in the first portion of the URL, e.g.
`/es/...`. Locale negotiation is done with jed/locale. The set of
available locales is generated by reading the `po` directory. This
means that if you want English as a supported locale, you must have
a `po/en.po` file.

This commit also splits out the route definitions into a separate
file so that they can be included multiple times in the middleware
chain.