Introducing Frable: a table for Fracas.
None of the existing Angular table/grids do what we need (n x m for one),
so we're building our own. Frable is based on ngTable. Actually, right now it
basically *is* ngTable renamed and with a few patches, but that should
hopefully change soon.

ngGrid was also evaluated. It has some more features than ngTable, but
it still doesn't do everything we need. It also uses divs for layout and
reminds me too much of ExtJS. ngTable is a lot simpler and is a better
base to build our own magic pivot-table-thing.