Reusable edit form
This commit introduces an `outpatientForm` directive that can be used
wherever you want to stick a form. To illustrate this, we added edit
controls to the workbench's table visualization.

This commit also changes the existing form to use a vertical layout
instead of a horizontal layout. This is in line with Bootstrap's default,
is less markup, and looks a lot better (no alignment issues).