Edit visit view
This commit introduces the "edit" view. It's currently a grid that displays
the existing records in the datastore, plus a button to add a new record.
In the future, we'll add filtering to this view so that users can find
specific records to edit.

This code is probably a little over-engineered because I thought
components like the table would be re-used. If that doesn't happen, we
should probably strip out all the event spaghetti.