Package JS libs in separate bundle
This is the de facto standard for how large SPAs are shipped since it
helps with caching. But more importantly for us, it makes debugging
a lot quicker since the browser doesn't have to parse a giant source map.

Note that there is still an outstanding issue with bundling d3, so that's
still included in the main bundle.