{{}} -> [[]] for EJS tags
`{{` is used by Angular. If everything goes right it shouldn't matter,
since the EJS tags will be parsed on the server before Angular gets
involved. But it was a little confusing that the tags looked the same,
and IntelliJ finally decided to interpret them as Angular expressions
in the latest update, causing all sorts of red squiggles.