Start versioning our datastore
This commit introduces a primitive mechanism for versioning our
elasticsearch instances. It's similar to Rails migrations, except that
we don't support the migration part (yet). But we do have the schema
and seeds part down. So for now the workflow is to destroy and recreate
the entire dataset. Luckily, we only have a few thousand records.