Local users
This commit introduces local users. Users are stored in an elasticsearch
index. Their passwords are hashed with scrypt.

Note that this commit does not implement role-based access, although it
lays some of the groundwork for it. More work needs to be done to move
the access control logic from the models into the controllers.