Massive codex refactoring
This commit introduces callbacks for controller classes to perform access
control. It also splits out `model.indices` into a separate class, since
the current implementation was broken and caused my tests to fail. This
means `model.index` is *not* a string anymore! It is an instance of
`Index` which has a `name` field.