Set <base> relative to actually URL clients hit
We can't set the base href relative to req since we might be behind a
reverse proxy. Note that this means you can only hit Fracas at the URL
specified by `conf.url`. That's probably OK since you can always
implement VHosts at the proxy level, but we may need to re-evaluate in
the future, e.g. if we have HTTP and HTTPS versions, or want to hit the
site via IP and DNS at the same time.