Per-request Model.get cache
Lots of preInsert and preDelete handlers need access to the underlying
document, e.g. for merging it with the client's request or performing
access control. This commit introduces a cache so that multiple handlers on
the same request do not need to duplicate the request. Thanks to
elasticsearch's versioning, we don't even need to worry about cache
invalidation.