Clustering
A single node process should be able to handle the low volume of requests
we expect, but clustering does provide support for restarting workers when
they fail. This way, an uncaught error in a single request doesn't take down
the whole server. It also provides some breathing room if processing a request
blocks for too long.