Put Codex model properties under `.doc`
Previously, Codex stored document properties, i.e. the document's
`_source`, directly on the model instance, while other model properties
were stored under a `_` object. This way, there is a separation between
properties that should be written to elasticsearch and those that shouldn't.
Unfortunately, this breaks down in the presence of 3rd-party code that doesn't
know about our conventions. For example, `EventEmitter` creates a `_events`
property that we almost certainly don't want written to elasticsearch.

To avoid all these headaches, this commit moves properties that belong in
elasticsearch to `.doc`. This is similar to what Sequelize does.