Docker
Docker is awesome. Unfortunately, running it in production is not as
easy as we need it to be. It still requires a lot of devops glue to
keep containers up to date. We don't care that much about HA, but we do
need maintenance costs to be as close to 0 as possible. So for now a
VM with unattended-upgrades makes more sense.