Reusable VM with Vagrant, CoreOS, and Docker
boot2docker is nice in that it's small and fast, but that's because
it runs entirely from memory. This means any VM would need to either
mount the ISO (which VirtualBox doesn't like to export) or install
the ISO to virtual disk. But at that point, why not just use CoreOS,
which is purpose built for Docker. So that's what we did. (Raw Ubuntu
with Docker, optionally installed via Vagrant's Docker provisioner,
had horrible performance and often timed out building images.)

To build an appliance, run `vagrant up`. In a few minutes, you should
have a working CoreOS VM with Docker containers for each component of
our stack. Note that ports 9000 (Fracas) and 9200 (elasticsearch) are
forwarded from the host to the guest. This means you can seed your data
before building an appliance.

Once your VM is ready to be exported, power down the VM and generate an
OVA file in VirtualBox ("Export Appliance"). The OVA is ~1GB, although
we can get it smaller if we need to.