Date of birth in addition to age
Lots of standards require collecting date of birth, in addition to
(or instead of) age. This commit adds some fancy validation so that
collecting date of birth and age isn't confusing. Basically, if date of
birth is set, than age is automatically calculated and inputted on the
form. However, the user can still tweak the age. This is useful for
rounding, e.g. going from an age of 0.99 to 1. If the user inputs an age
that is way off though, like age=1 and date of birth=01/01/1970, than we
display a warning.