Split age into age.years and age.months
Before, `patient.age` was a floating point number, so users could input
fractional ages. That works fine for things like six months (=0.5), but
not so much for things like one month (=0.083333). So this commit splits
age into the pair of rational numbers that it is.

Note that we could track intervals smaller than a month, e.g. weeks,
days, etc. But there's no pressing need for that right now.