Proof of concept: analysis on categorical fields
We can already collect any geography (or anything else) on facility.
This commit always you to do analysis on those fields, e.g. use
"district" as a filter. It does this by (ab)using the fact that,
although these fields are not visible to the user on data entry, they
are being collected. For example, location fields--like district--are
not entered by the user on data entry. Instead, the user selects a
facility, which has a location. So it makes sense that location is an
"enabled" field from the perspective of the form.

Another way to put it: the form config specifies what data is being
collected, not just what fields to display.