VM: load services from fracas-services image
CoreOS is awesome in that the entire configuration of the OS is in a few
short systemd unit files. But how do those files get onto the machine?
And how are they managed, e.g. how do they get updated?

Answer: stick them in a container! This way, updates can be fetched with
a simple `docker pull`. See the Vagrantfile for how to get the services
out of the container and onto the host, and for what to do after they're
on the host.